Properly configure EF Keyless entity
Before there was a bug where a migration would be added for the keyless entity, so we needed that "hack" on the call Toview. With the latest version of EF, calling ToView is expected for the object to exist in the DB and since it doesn't in this case, it fails at runtime with System.ArgumentNullException: Value cannot be null. (Parameter 'clrName').